Extract Parser out of Graph
