+++ Makefile
@@ -1,7 +1,7 @@
 CXX=g++
 CXXFLAGS=-std=c++11 -m64 -g -D__STDC_FORMAT_MACROS -O0 -c -Wall
 LDFLAGS=-m64 -g -ljansson -lreadline -lcityhash
-SOURCES=main.cc ruby_heap_obj.cc graph.cc dominator_tree.cc progress.cc
+SOURCES=main.cc ruby_heap_obj.cc parser.cc graph.cc dominator_tree.cc progress.cc
 OBJECTS=$(SOURCES:.cc=.o)
 EXECUTABLE=harb
